phase-II/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend"
version = "0.1.0"
description = "Backend for the Todo Full-Stack Web Application"
authors = [
    {name = "Developer", email = "developer@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "sqlmodel>=0.0.16",
    "asyncpg>=0.29.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.1",
    "sqlalchemy>=2.0.23"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"